@@ -2,10 +2,10 @@
  * Contains the miscellaneous route handlers.
  */
 class AppController {
-    static getHomepage(request, response) {
-      response.status(200).send('Hello Holberton School!');
-    }
+  static getHomepage(request, response) {
+    response.status(200).send('Hello Holberton School!');
   }
+}
   
-  export default AppController;
-  module.exports = AppController;
+export default AppController;
+module.exports = AppController;
